move long constants to their own file
